OJ-2279: Don't use the stack name in state machines names
The state machine names have a shorter length restriction that stack names, so if the stack name is too long the deployment will fail. Instead, let CloudFormation generate unique names, which by default include the stack information.

Use a reference to the state machine resource to get its ARN. This ensures the ARN is always correct even if the name of the state machine changes.